backend/requirements.txt
# Backend Dependencies for GitHub Issue Assistant
# Pinned versions for reproducibility

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data Validation
pydantic==2.5.3

# Environment Variables
python-dotenv==1.0.0

# HTTP Client (Async)
httpx==0.26.0

# Google Gemini AI (updated for current model support)
google-generativeai>=0.8.0

# Typing extensions (for Python 3.11 compatibility)
typing-extensions==4.9.0
